fix zrange

+++ src/helper.rs
@@ -3,7 +3,7 @@ use self::time::Timespec;
 use self::time::at;
 use self::time::strftime;
 
-pub fn ts2str(ts: u32) -> String {
+pub fn ts2str(ts: isize) -> String {
     let timespec = Timespec::new(1433860335, 0);
     let tm = at(timespec);
     let result = strftime("%F %T", &tm).unwrap();
